refactor(js): split templates module
